+++ ForestObj.h
@@ -45,7 +45,7 @@ public:
     bool allowNew();
     void load(int x, int y);
     void set(int sh, FileBuffer* data);
-    void set(QString sh, int val);
+    void set(QString sh, long long int val);
     void set(QString sh, float val);
     void set(QString sh, FileBuffer* data);
     void save(QTextStream* out);
