+++ SignalObj.h
@@ -23,7 +23,7 @@ public:
     SignalObj(const SignalObj& orig);
     virtual ~SignalObj();
     void load(int x, int y);
-    void set(QString sh, int val);
+    void set(QString sh, long long int val);
     void set(int sh, FileBuffer* val);
     void set(QString sh, QString val);
     void set(QString sh, FileBuffer* data);
