+++ SoundRegionObj.h
@@ -27,7 +27,7 @@ public:
     void initTrItems(float* tpos);
     void deleteTrItems();
     void load(int x, int y);
-    void set(QString sh, int val);
+    void set(QString sh, long long int val);
     void set(int sh, FileBuffer* val);
     void set(QString sh, QString val);
     void set(QString sh, FileBuffer* data);
