+++ SoundSourceObj.cpp
@@ -41,7 +41,7 @@ void SoundSourceObj::load(int x, int y) {
     setMartix();
 }
 
-void SoundSourceObj::set(QString sh, int val){
+void SoundSourceObj::set(QString sh, long long int val){
     if (sh == ("ref_value")) {
         this->fileName = "";
         for (auto it = Game::soundList->sources.begin(); it != Game::soundList->sources.end(); ++it ){
