+++ SpeedpostObj.h
@@ -25,7 +25,7 @@ public:
     SpeedpostObj(const SpeedpostObj& orig);
     virtual ~SpeedpostObj();
     void load(int x, int y);
-    void set(QString sh, int val);
+    void set(QString sh, long long int val);
     void set(int sh, FileBuffer* val);
     void set(QString sh, QString val);
     void set(QString sh, FileBuffer* data);
